Update and rename NetSalary.c to NetSalary.cpp
fix file type error and, error in code
